# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SRA Resource Management System API
  version: 1.0.0
  description: >
    Business-layer API for the SRA Resource Management System (SRA-RMS).
    Provides CRUD over Clients, Projects, Resources, and Allocations, plus
    dashboard and reporting endpoints. Authentication is delegated to
    Microsoft Active Directory (OAuth2 / OpenID Connect); authorization is
    role-based (Administrator, General, Report).
  contact:
    name: SRA
    email: umesh.kodippili@sra.com.au
  license:
    name: Proprietary - SRA
servers:
  - url: https://api.rms.sra.com.au/v1
    description: Production
  - url: https://api.staging.rms.sra.com.au/v1
    description: Staging
  - url: http://localhost:5000/v1
    description: Local development
security:
  - oauth2: []
...
paths:
...
  /resources:
    get:
      tags: [Resources]
      summary: List resources
      operationId: listResources
      description: Roles - Administrator, General. Supports skill/department/location/status filters.
      parameters:
        - $ref: '#/components/parameters/Search'
        - name: skill
          in: query
          description: Filter by one or more skill tags (AND semantics).
          schema:
            type: array
            items: { type: string }
          style: form
          explode: true
        - name: department
          in: query
          schema: { type: string }
        - name: location
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ResourceStatus' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A paged list of resources.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResourcePage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Resources]
      summary: Create a resource
      operationId: createResource
      description: Roles - Administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceCreate' }
      responses:
        '201':
          description: Resource created.
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resource' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /resources/{resourceId}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Resources]
      summary: Get a resource
      operationId: getResource
      description: Roles - Administrator, General. Includes current and historical allocations.
      responses:
        '200':
          description: The resource.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResourceDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Resources]
      summary: Update a resource
      operationId: updateResource
      description: Roles - Administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceUpdate' }
      responses:
        '200':
          description: Resource updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resource' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Resources]
      summary: Delete a resource
      operationId: deleteResource
      description: >
        Roles - Administrator. Fails with 409 if the resource has allocations,
        unless cascade=true is supplied.
      parameters:
        - $ref: '#/components/parameters/Cascade'
      responses:
        '204': { description: Resource deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /resources/{resourceId}/image:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    put:
      tags: [Resources]
      summary: Upload a resource profile image
      operationId: uploadResourceImage
      description: Roles - Administrator.
      requestBody:
        required: true
        content:
          image/png:
            schema: { type: string, format: binary }
          image/jpeg:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: Image stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resource' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /resources/{resourceId}/allocations:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Resources]
      summary: List a resource's allocations
      operationId: listResourceAllocations
      description: Roles - Administrator, General.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: A paged list of the resource's allocations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------- Allocations ----------------
  /allocations:
    get:
      tags: [Allocations]
      summary: List allocations
      operationId: listAllocations
      description: Roles - Administrator, General. Filter by project, resource, or date range.
      parameters:
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
        - name: resourceId
          in: query
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A paged list of allocations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Allocations]
      summary: Create an allocation
      operationId: createAllocation
      description: >
        Roles - Administrator. projectId and resourceId are supplied in the body.
        Validates dates against the project window; over-allocation returns
        warnings on the created object.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllocationCreateFull' }
      responses:
        '201':
          description: Allocation created.
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Allocation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: Microsoft Active Directory / Entra ID (OpenID Connect).
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/common/oauth2/v2.0/token
          scopes:
            api://sra-rms/access_as_user: Access the SRA-RMS API as the signed-in user.
  headers:
    Location:
      description: URI of the newly created resource.
      schema: { type: string, format: uri }
  parameters:
    Search:
      name: q
      in: query
      description: Free-text search term.
      schema: { type: string }
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: Items per page.
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    Sort:
      name: sort
      in: query
      description: 'Sort expression, e.g. "name" or "-startDate" (prefix - for descending).'
      schema: { type: string }
    Cascade:
      name: cascade
      in: query
      description: If true, delete dependents instead of returning 409.
      schema: { type: boolean, default: false }
...
    ResourceId:
      name: resourceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The request was invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication is required or has failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The authenticated user lacks the required role.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: The request conflicts with the current state (e.g. duplicate or dependents exist).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }

    # ---- Enums ----
...
    ResourceStatus:
      type: string
      enum: [active, inactive, onLeave]
    DayOfWeek:
      type: string
      enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
    EffortUnit:
      type: string
      enum: [hoursPerWeek, percent]

    # ---- Client ----
...
    ResourceBase:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        email: { type: string, format: email }
        primaryJobTitle: { type: string }
        secondaryJobTitle: { type: string }
        status: { $ref: '#/components/schemas/ResourceStatus' }
        department: { type: string }
        location: { type: string }
        notes: { type: string }
        skills:
          type: array
          items: { type: string }
        availabilityHoursPerWeek: { type: number, format: double, minimum: 0, maximum: 168 }
        workingDays:
          type: array
          items: { $ref: '#/components/schemas/DayOfWeek' }
    ResourceCreate:
      allOf:
        - $ref: '#/components/schemas/ResourceBase'
        - type: object
          required: [name, email, primaryJobTitle, availabilityHoursPerWeek]
    ResourceUpdate:
      allOf:
        - $ref: '#/components/schemas/ResourceBase'
        - type: object
          required: [name, email, primaryJobTitle, availabilityHoursPerWeek]
    Resource:
      allOf:
        - $ref: '#/components/schemas/ResourceBase'
        - type: object
          required: [id, name, email]
          properties:
            id: { type: string, format: uuid }
            imageUrl: { type: string, format: uri }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    ResourceDetail:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            allocations:
              type: array
              items: { $ref: '#/components/schemas/Allocation' }
            allocatedHoursPerWeek:
              type: number
              format: double
              description: Sum of current allocations' effort in hours/week.
    ResourcePage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Resource' }
        meta: { $ref: '#/components/schemas/PageMeta' }

    # ---- Allocation ----
    AllocationBase:
      type: object
      properties:
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        effort: { type: number, format: double, minimum: 0 }
        effortUnit: { $ref: '#/components/schemas/EffortUnit' }
        roleOnProject: { type: string }
        billable: { type: boolean }
...
    AllocationCreateFull:
      description: Used on the top-level /allocations endpoint.
      allOf:
        - $ref: '#/components/schemas/AllocationBase'
        - type: object
          required: [projectId, resourceId, startDate, endDate, effort, effortUnit]
          properties:
            projectId: { type: string, format: uuid }
            resourceId: { type: string, format: uuid }
...
    Allocation:
      allOf:
        - $ref: '#/components/schemas/AllocationBase'
        - type: object
          required: [id, projectId, resourceId, startDate, endDate, effort, effortUnit]
          properties:
            id: { type: string, format: uuid }
            projectId: { type: string, format: uuid }
            projectName: { type: string }
            resourceId: { type: string, format: uuid }
            resourceName: { type: string }
            warnings:
              type: array
              description: Non-blocking validation warnings, e.g. over-allocation.
              items: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    AllocationPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Allocation' }
        meta: { $ref: '#/components/schemas/PageMeta' }

    # ---- Dashboard ----
...